Enable output like tee
Enables standard output and standard error output like tee.
